Issue-148 - activityType searching call proto added, AsNoTrucking add… (#152)
* Issue-148 - activityType searching call proto added, AsNoTrucking added in repository's get calls

* Issue-148 - space,  AsEnumerable removed

+++ src/TourManager/TourManager.Common/Clients/ActivityType.cs
@@ -13,6 +13,6 @@ namespace TourManager.Common.Clients
         /// </summary>
         public long? Id { get; set; }
         public string ActivityName { get; set; }
-        public int AgencyId { get; set; }
+        public long AgencyId { get; set; }
     }
 }
